requirements.txt
aioredis==2.0.1
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.1
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.2
exceptiongroup==1.2.2
fastapi==0.113.0
fastapi-cache2==0.2.2
fastapi-cli==0.0.5
fastapi-users==13.0.0
flower==2.0.1
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
humanize==4.10.0
idna==3.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jose==1.0.0
kombu==5.4.0
makefun==1.15.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.7
packaging==24.1
passlib==1.7.4
pendulum==3.0.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pwdlib==0.2.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.9.0
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.23.2
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.3.2
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
rich==13.8.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.4
time-machine==2.15.0
tomli==2.0.1
tornado==6.4.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.6
uvloop==0.20.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
